robo-nexus-birthday-bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
hypothesis>=6.82.0
pytest>=7.4.0
pytest-asyncio>=0.21.0